+++ src/Krystal/Http/CookieBag.php
@@ -12,6 +12,7 @@
 namespace Krystal\Http;
 
 use Krystal\Date\TimeHelper;
+use Krystal\Security\CrypterInterface;
 use RuntimeException;
 use InvalidArgumentException;
 use UnexpectedValueException;
@@ -26,14 +27,38 @@ final class CookieBag implements CookieBagInterface, PersistentStorageInterface
     private $cookies = array();
 
     /**
+     * Any compliant crypter implementation
+     * 
+     * @var \Krystal\Security\CrypterInterface
+     */
+    private $crypter;
+
+    /**
      * State initialization
      * 
      * @param array $cookie
+     * @param \Krystal\Security\CrypterInterface $crypter
      * @return void
      */
-    public function __construct(array &$cookie)
+    public function __construct(array &$cookie, CrypterInterface $crypter = null)
     {
         $this->cookies  = &$cookie;
+        $this->crypter = $crypter;
+    }
+
+    /**
+     * Returns a crypter
+     * 
+     * @throws \RuntimeException If the instance wasn't provided
+     * @return \Krystal\Security\CrypterInterface
+     */
+    private function getCrypter()
+    {
+        if ($this->crypter instanceof CrypterInterface) {
+            return $this->crypter;
+        } else {
+            throw new RuntimeException('An instance of crypter was not provided. Therefore can not work with encryption and decryption');
+        }
     }
 
     /**
@@ -118,6 +143,25 @@ final class CookieBag implements CookieBagInterface, PersistentStorageInterface
     }
 
     /**
+     * Sets a encrypted cookie
+     * 
+     * @param string $key Cookie key
+     * @param string $value Cookie value
+     * @param integer $ttl Cookie time to live in seconds
+     * @param string $path The path on the server in which the cookie will be available on. Defaults to /
+     * @param boolean $secure Indicates that the cookie should only be transmitted over a secure HTTPS connection from the client
+     * @param boolean $httpOnly When TRUE the cookie will be made accessible only through the HTTP protocol
+     * @param boolean $raw Whether to send a cookie without urlencoding the cookie value
+     * @throws \InvalidArgumentException if either $key or $value isn't scalar by type
+     * @throws \UnexpectedValueException If trying to set a key that contains a dot
+     * @return boolean
+     */
+    public function setEncrypted($key, $value, $ttl = TimeHelper::YEAR, $path = '/', $secure = false, $httpOnly = false, $raw = false)
+    {
+        return $this->set($key, $this->getCrypter()->encryptValue($value), $ttl, $path, $secure, $httpOnly, $raw);
+    }
+
+    /**
      * Retrieves a key form cookies
      * 
      * @param string $key
@@ -133,6 +177,18 @@ final class CookieBag implements CookieBagInterface, PersistentStorageInterface
                 'Attempted to read non-existing cookie "%s"', $key
             ));
         }
+    }
+
+    /**
+     * Retrieves a key form cookies decoding its value
+     * 
+     * @param string $key
+     * @throws \RuntimeException if $key does not exist in cookies`
+     * @return string
+     */
+    public function getEncrypted($key)
+    {
+        return $this->getCrypter()->decryptValue($this->get($key));
     }
 
     /**